+++ contracts/handlers/aave/HAaveProtocol.sol
@@ -3,7 +3,9 @@ pragma experimental ABIEncoderV2;
 
 import "@openzeppelin/contracts/token/ERC20/SafeERC20.sol";
 import "@openzeppelin/contracts/math/SafeMath.sol";
+import "./IAToken.sol";
 import "./ILendingPool.sol";
+import "./ILendingPoolCore.sol";
 import "./ILendingPoolAddressesProvider.sol";
 import "./FlashLoanReceiverBase.sol";
 import "../HandlerBase.sol";
@@ -13,9 +15,10 @@ interface IProxy {
     function execs(address[] calldata tos, bytes[] calldata datas) external;
 }
 
-
 contract HAaveProtocol is HandlerBase, FlashLoanReceiverBase {
     using SafeERC20 for IERC20;
+    
+    uint16 constant REFERRAL_CODE = 56;
 
     function flashLoan(
         address _reserve,
@@ -43,5 +46,45 @@ contract HAaveProtocol is HandlerBase, FlashLoanReceiverBase {
         );
         IProxy(address(this)).execs(tos, datas);
         transferFundsBackToPoolInternal(_reserve, _amount.add(_fee));
+    }
+
+    function deposit(
+        address _reserve,
+        uint256 _amount
+    ) external payable {
+        ILendingPool lendingPool = ILendingPool(
+            ILendingPoolAddressesProvider(PROVIDER).getLendingPool()
+        );
+
+        if (_reserve == ETHADDRESS) {
+            lendingPool.deposit.value(_amount)(_reserve, _amount, REFERRAL_CODE);
+        } else {
+            address lendingPoolCore = ILendingPoolAddressesProvider(PROVIDER).getLendingPoolCore();
+            IERC20(_reserve).safeApprove(lendingPoolCore, _amount);
+            lendingPool.deposit(_reserve, _amount, REFERRAL_CODE);
+            IERC20(_reserve).safeApprove(lendingPoolCore, 0);
+        }
+        
+        address aToken = _getAToken(_reserve);
+        _updateToken(aToken);
+    }
+
+    function redeem(
+        address _aToken,
+        uint256 _amount
+    ) external payable {
+        IAToken(_aToken).redeem(_amount);
+        address underlyingAsset = IAToken(_aToken).underlyingAssetAddress();
+        if (underlyingAsset != ETHADDRESS)  _updateToken(underlyingAsset);
+    }
+
+    function _getAToken(address _reserve) internal view returns(address) {
+        ILendingPoolCore lendingPoolCore = ILendingPoolCore(
+            ILendingPoolAddressesProvider(PROVIDER).getLendingPoolCore()
+        );
+        address aToken = lendingPoolCore.getReserveATokenAddress(_reserve);
+        require(aToken != address(0), "aToken should not be zero address");
+
+        return aToken;
     }
 }